Run tests against an AllJoyn router.
There is now a packages.json file with right dependencies
defined so that the tests can now be run with "npm test".

The test runner scripts is in tests/run.js and it takes care
of starting an AllJoyn router so that also the tests that rely
on router connection can be run.