Don't define the function within a for-loop.
Doing this is against the coding conventions we now have in place.

At the same time, improved the order of function calls so that there
would be less change that problems would occur in certain conditions.

A few console logs were added for error scenarios that are not supported,
but can be met if the plugin API is called with unsupported arguments.