Add Internet (Client & Server) capability.
This capability is required in the Windows Store apps for AllJoyn connections
to work.

This capability should be automatically added if app is run from Visual Studio,
but it seems like this is not the case at least always. This commit is at least
required when app is run with Cordova scripts.

Addresses issue #14.